+++ includes/libraries/class-libraries.php
@@ -132,18 +132,18 @@ class Libraries {
 			'license' => 'mit',
 			'langs'   => 'en',
 		];
-		self::$mono_libraries              = [];
-		self::$mono_libraries['parsedown'] = [
-			'name'    => 'Parsedown',
-			'detect'  => 'Parsedown',
-			'base'    => DECALOG_VENDOR_DIR . 'parsedown/',
-			'version' => '1.8.0-beta-7',
+		self::$psr4_libraries['markdown'] = [
+			'name'    => 'Markdown Parser',
+			'prefix'  => 'cebe\markdownparser',
+			'base'    => DECALOG_VENDOR_DIR . 'markdown/',
+			'version' => '1.2.1',
 			// phpcs:ignore
-			'author'  => sprintf( esc_html__( '%s & contributors', 'decalog' ), 'Emanuil Rusev' ),
-			'url'     => 'https://parsedown.org',
+			'author'  => sprintf( esc_html__( '%s & contributors', 'decalog' ), 'Carsten Brandt' ),
+			'url'     => 'https://github.com/cebe/markdown',
 			'license' => 'mit',
 			'langs'   => 'en',
 		];
+		self::$mono_libraries             = [];
 	}
 
 	/**